+++ src/main/java/edu/cqu/no1/action/ExpertTeamAction.java
@@ -1180,7 +1180,7 @@ public class ExpertTeamAction extends BaseAction {
             teacher.setTeaCode(teacherCode);
             //
             if (teacherAge != null && !teacherAge.trim().equals("")) {
-                teacher.setTeaAge(new BigDecimal(teacherAge));
+                teacher.setTeaAge(new Double(teacherAge));
             }
             teacher.setTeaSex(teacherSex);
             teacher.setTeaTitle(teacherTitle);
